CXP-791: refactor GetWebMarketplaceUrl endpoint
