+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/InternalApi/Controller/Marketplace/GetWebMarketplaceUrl.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller;
+namespace Akeneo\Connectivity\Connection\Infrastructure\InternalApi\Controller\Marketplace;
 
 use Akeneo\Connectivity\Connection\Domain\Marketplace\MarketplaceUrlGeneratorInterface;
 use Akeneo\UserManagement\Bundle\Context\UserContext;
@@ -16,7 +16,7 @@ use Symfony\Component\HttpFoundation\Response;
  * @copyright 2021 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class MarketplaceUrlController
+class GetWebMarketplaceUrl
 {
     private MarketplaceUrlGeneratorInterface $marketplaceUrlGenerator;
     private UserContext $userContext;
@@ -29,7 +29,7 @@ class MarketplaceUrlController
         $this->userContext = $userContext;
     }
 
-    public function get(Request $request): Response
+    public function __invoke(Request $request): Response
     {
         if (!$request->isXmlHttpRequest()) {
             return new RedirectResponse('/');
