+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/routing/marketplace.yml
@@ -1,3 +1,8 @@
+akeneo_connectivity_connection_marketplace_rest_get_web_marketplace_url:
+    path: '/marketplace-url'
+    controller: akeneo_connectivity.connection.internal_api.controller.marketplace.get_web_marketplace_url
+    methods: [GET]
+
 akeneo_connectivity_connection_marketplace_rest_get_all_extensions:
     path: '/extensions'
     controller: akeneo_connectivity.connection.internal_api.controller.marketplace.get_all_extensions
