+++ src/Akeneo/Connectivity/Connection/front/src/connect/hooks/use-fetch-marketplace-url.ts
@@ -2,7 +2,7 @@ import {useCallback} from 'react';
 import {useRoute} from '../../shared/router';
 
 export const useFetchMarketplaceUrl = () => {
-    const url = useRoute('akeneo_connectivity_connection_marketplace_url_rest_get');
+    const url = useRoute('akeneo_connectivity_connection_marketplace_rest_get_web_marketplace_url');
 
     return useCallback(async () => {
         const response = await fetch(url, {
